requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
pdfplumber>=0.10.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
Pillow>=10.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
requests>=2.31.0

